package/pyproject.toml
[tool.poetry]
name = "healthcareai_toolkit"
version = "0.1.0"
description = "A toolkit for healthcare AI examples, providing various utilities and models to streamline the use of the examples in Microsoft HealthcareAI ecosystem."
authors = ["Healthcare AI Team <medimagingaihls@microsoft.com>"]
packages = [
    {include = "healthcareai_toolkit", from = "."},
    {include = "model_library", from = "."}
]

[tool.poetry.dependencies]
python = ">=3.9.0,<3.12"
azure-ai-ml = "~1.20.0"
azure-storage-file-datalake = "~12.17.0"
azureml-fsspec = "~1.3.0"
nbformat = "~5.10.0"
plotly = "~5.23.0"
scikit-learn = "~1.5.0"
torch = "~2.4.0"
torchvision = "~0.19.0"
tqdm = "~4.66.5"
python-gdcm = "~3.0.24.0"
gdown = "~5.2.0"
SimpleITK = "~2.4.0"
opencv-python = "~4.10.0.84"
pydicom = "~2.4.0"
azure-identity = "~1.19.0"
timm = "~1.0.10"
transformers = "~4.16.2"
setuptools = "~59.8.0"
einops = "~0.8.0"
mup = "~1.0.0"
fvcore = "~0.1.5.post20221221"
sentencepiece = "~0.2.0"
tenacity = "~9.0.0"
ftfy = "~6.2.3"
ipykernel = "~6.29.5"
pandas = "~2.0.3"
jupyter = "~1.1.1"
pillow = "~10.4.0"
matplotlib = "~3.7.5"
numpy = "~1.24.4"
openai = "~1.51.2"
umap-learn = "~0.5.6"
scipy = "~1.10.1"
azureml-core = "~1.57.0.post3"
ratelimit = "~2.2.1"
python-magic = "~0.4.27"
scikit-image = "~0.24.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"